Update README.md
Working with the do-while Loop

Compile and run the program several times with various inputs. Try all the possible relevant cases.

What do you think will happen if you do not enter A, B, C, D or E? Try running the program and inputting another letter. Add comments to the code where the program handles these invalid inputs.

Replace the line
if (validBeverage == true)
with the line
if (validBeverage)
and run the program again. Are there any differences in the execution of the program? Add comments to the bottom of the program explaining why or why not.